# internal/interface/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Blog API
  version: 1.0.0
  description: Clean Architecture starter kit blog API.
servers:
  - url: http://localhost:8080
paths:
  /health/live:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: OK
  /health/ready:
    get:
      summary: Readiness probe
      responses:
        "200":
          description: Ready
  /openapi.yaml:
    get:
      summary: Raw OpenAPI specification
      responses:
        "200":
          description: OK
  /docs:
    get:
      summary: Interactive API documentation
      responses:
        "200":
          description: OK
  /api/v1/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/v1/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest'
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/refresh:
    post:
      summary: Rotate refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        "200":
          description: Rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/logout:
    post:
      summary: Logout one session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        "204":
          description: Logged out
  /api/v1/auth/logout-all:
    post:
      summary: Logout all sessions
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out from all sessions
  /api/v1/auth/sessions:
    get:
      summary: List active and historical sessions for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
  /api/v1/auth/sessions/revoke:
    post:
      summary: Revoke one session by ID
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeSessionRequest'
      responses:
        "204":
          description: Revoked
  /api/v1/posts:
    get:
      summary: List posts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPostResponse'
    post:
      summary: Create post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
  /api/v1/posts/{postID}:
    put:
      summary: Update post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
    delete:
      summary: Delete post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postID
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /api/v1/posts/{postID}/likes:
    post:
      summary: Like post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postID
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Liked
  /api/v1/posts/{postID}/comments:
    get:
      summary: List comments for post
      parameters:
        - in: path
          name: postID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCommentResponse'
    post:
      summary: Add comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommentRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
  /api/v1/comments/{commentID}:
    put:
      summary: Update comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: commentID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
    delete:
      summary: Delete comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: commentID
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /api/v1/admin/users:
    get:
      summary: List users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserResponse'
  /api/v1/admin/users/{userID}/role:
    put:
      summary: Update user role
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterUserRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginUserRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        device_name:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    RevokeSessionRequest:
      type: object
      required: [session_id]
      properties:
        session_id:
          type: string
    CreatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        publish_now:
          type: boolean
    UpdatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
    AddCommentRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
    UpdateCommentRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
    UpdateUserRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [user, admin]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        access_token:
          type: string
        refresh_token:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        created_at:
          type: string
          format: date-time
    PostResponse:
      type: object
      properties:
        id:
          type: string
        author_id:
          type: string
        title:
          type: string
        content:
          type: string
        status:
          type: string
        likes_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CommentResponse:
      type: object
      properties:
        id:
          type: string
        post_id:
          type: string
        author_id:
          type: string
        body:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SessionResponse:
      type: object
      properties:
        id:
          type: string
        device_name:
          type: string
        user_agent:
          type: string
        ip_address:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    PageMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
    PaginatedPostResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PostResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'
    PaginatedCommentResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'
    PaginatedUserResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'